Refactor triangle mesh.
